refactor(kimi-k2): split weights module root
