+++ pegainfer-kimi-k2/src/weights/package.rs
@@ -1,4 +1,147 @@
 use super::*;
+use super::{
+    load::dtype_element_bytes,
+    manifest::{
+        KimiInt4ProjectionWeightNames, KimiLayerWeightKindNames, KimiMoeLayerWeightNames,
+        KimiRankWeightNames, KimiRoutedExpertWeightNames, validate_rank_tensor_catalog,
+    },
+};
+
+pub(crate) struct KimiGpuRawTensor {
+    pub name: String,
+    pub dtype: Dtype,
+    pub shape: Vec<usize>,
+    pub bytes: usize,
+    pub data: CudaSlice<u8>,
+}
+
+pub(crate) struct KimiRankGpuWeights {
+    pub rank: usize,
+    pub tensors: BTreeMap<String, KimiGpuRawTensor>,
+    pub total_bytes: usize,
+}
+
+pub(crate) struct KimiRouterGpuWeights<'a> {
+    pub gate_weight: &'a KimiGpuRawTensor,
+    pub e_score_correction_bias: &'a KimiGpuRawTensor,
+}
+
+pub(crate) struct KimiRouterDeviceWeights {
+    pub gate_weight: GpuWeight<KIMI_K2_ROUTED_EXPERTS, KIMI_K2_HIDDEN>,
+    pub e_score_correction_bias: CudaSlice<f32>,
+}
+
+struct KimiInt4ProjectionGpuWeights<'a> {
+    pub weight_packed: &'a KimiGpuRawTensor,
+    pub weight_scale: &'a KimiGpuRawTensor,
+    pub weight_shape: &'a KimiGpuRawTensor,
+}
+
+#[derive(Clone, Copy, Debug, Eq, PartialEq)]
+pub(crate) enum KimiInt4ProjectionRole {
+    Gate,
+    Up,
+    Down,
+}
+
+impl KimiInt4ProjectionRole {
+    const fn dims(self) -> (usize, usize) {
+        match self {
+            Self::Gate | Self::Up => (KIMI_K2_EXPERT_INTERMEDIATE, KIMI_K2_HIDDEN),
+            Self::Down => (KIMI_K2_HIDDEN, KIMI_K2_EXPERT_INTERMEDIATE),
+        }
+    }
+
+    const fn kernel_role(self) -> KimiInt4ExpertRole {
+        match self {
+            Self::Gate => KimiInt4ExpertRole::W1Gate,
+            Self::Up => KimiInt4ExpertRole::W3Up,
+            Self::Down => KimiInt4ExpertRole::W2Down,
+        }
+    }
+}
+
+#[derive(Clone, Debug, Eq, PartialEq)]
+pub(crate) struct KimiExpertMajorProjectionPlan {
+    pub local_experts: usize,
+    pub out_dim: usize,
+    pub in_dim: usize,
+    pub packed_bytes: usize,
+    pub scale_bytes: usize,
+}
+
+pub(crate) struct KimiExpertMajorProjectionMarlinBuffers {
+    pub role: KimiInt4ProjectionRole,
+    pub plan: KimiExpertMajorProjectionPlan,
+    pub manifest: KimiInt4WeightManifest,
+    pub weight_packed_marlin_uint4b8: CudaSlice<u8>,
+    pub weight_scale_marlin_permuted: CudaSlice<bf16>,
+}
+
+impl KimiExpertMajorProjectionMarlinBuffers {
+    fn as_marlin_weight(&self) -> KimiMarlinInt4Weight<'_> {
+        KimiMarlinInt4Weight {
+            manifest: self.manifest,
+            weight_packed_uint4b8: &self.weight_packed_marlin_uint4b8,
+            weight_scale_permuted: &self.weight_scale_marlin_permuted,
+        }
+    }
+
+    fn package_bytes(&self) -> usize {
+        self.weight_packed_marlin_uint4b8.len()
+            + self.weight_scale_marlin_permuted.len() * std::mem::size_of::<bf16>()
+    }
+}
+
+pub(crate) struct KimiExpertMajorW13MarlinBuffers {
+    pub local_experts: usize,
+    pub in_dim: usize,
+    pub intermediate_dim: usize,
+    pub group_size: usize,
+    pub weight_packed_marlin_uint4b8: CudaSlice<u8>,
+    pub weight_scale_marlin_permuted: CudaSlice<bf16>,
+}
+
+impl KimiExpertMajorW13MarlinBuffers {
+    fn as_marlin_weight(&self) -> KimiMarlinFusedW13Int4Weight<'_> {
+        KimiMarlinFusedW13Int4Weight {
+            local_experts: self.local_experts,
+            in_dim: self.in_dim,
+            intermediate_dim: self.intermediate_dim,
+            group_size: self.group_size,
+            weight_packed_uint4b8: &self.weight_packed_marlin_uint4b8,
+            weight_scale_permuted: &self.weight_scale_marlin_permuted,
+        }
+    }
+
+    fn package_bytes(&self) -> usize {
+        self.weight_packed_marlin_uint4b8.len()
+            + self.weight_scale_marlin_permuted.len() * std::mem::size_of::<bf16>()
+    }
+}
+
+pub(crate) struct KimiMoeLayerExpertMarlinWeights {
+    pub layer_idx: usize,
+    pub w13: KimiExpertMajorW13MarlinBuffers,
+    pub down: KimiExpertMajorProjectionMarlinBuffers,
+    pub total_bytes: usize,
+}
+
+pub(crate) struct KimiRankExpertMarlinWeights {
+    pub rank: usize,
+    pub local_expert_range: Range<usize>,
+    pub layers: Vec<KimiMoeLayerExpertMarlinWeights>,
+    pub total_bytes: usize,
+}
+
+impl KimiMoeLayerExpertMarlinWeights {
+    pub(crate) fn as_marlin_weights(&self) -> KimiMarlinInt4ExpertWeights<'_> {
+        KimiMarlinInt4ExpertWeights {
+            w13: self.w13.as_marlin_weight(),
+            w2_down: self.down.as_marlin_weight(),
+        }
+    }
+}
 
 impl KimiGpuRawTensor {
     pub(crate) fn copy_bf16_matrix(
@@ -325,16 +468,16 @@ fn pack_expert_major_projection_marlin_buffers<'a>(
     );
 
     let mut weight_packed_offset_binary = ctx
-        .stream
+        .stream()
         .alloc_zeros::<u8>(manifest.packed_shape.elements())?;
     let mut weight_packed_marlin_uint4b8 = ctx
-        .stream
+        .stream()
         .alloc_zeros::<u8>(manifest.packed_shape.elements())?;
     let mut weight_scale_checkpoint = ctx
-        .stream
+        .stream()
         .alloc_zeros::<bf16>(manifest.scale_shape.elements())?;
     let mut weight_scale_marlin_permuted = ctx
-        .stream
+        .stream()
         .alloc_zeros::<bf16>(manifest.scale_shape.elements())?;
 
     copy_projection_component_to_contiguous(
@@ -396,10 +539,10 @@ fn fuse_expert_major_w13_marlin_buffers(
         gate.plan,
         up.plan
     );
-    let mut weight_packed_marlin_uint4b8 = ctx.stream.alloc_zeros::<u8>(
+    let mut weight_packed_marlin_uint4b8 = ctx.stream().alloc_zeros::<u8>(
         gate.weight_packed_marlin_uint4b8.len() + up.weight_packed_marlin_uint4b8.len(),
     )?;
-    let mut weight_scale_marlin_permuted = ctx.stream.alloc_zeros::<bf16>(
+    let mut weight_scale_marlin_permuted = ctx.stream().alloc_zeros::<bf16>(
         gate.weight_scale_marlin_permuted.len() + up.weight_scale_marlin_permuted.len(),
     )?;
 
@@ -443,7 +586,7 @@ fn copy_projection_component_to_contiguous<'a>(
             end <= expected_bytes,
             "Kimi expert-major {component} copy would exceed destination: end {end}, expected {expected_bytes}"
         );
-        ctx.stream
+        ctx.stream()
             .memcpy_dtod(
                 &tensor.data.slice(0..tensor.bytes),
                 &mut dst.slice_mut(offset..end),
@@ -475,12 +618,14 @@ fn copy_raw_tensor_to_typed<T: DeviceRepr + ValidAsZeroBits>(
         tensor.bytes,
         element_bytes
     );
-    let mut dst = ctx.stream.alloc_zeros::<T>(tensor.bytes / element_bytes)?;
+    let mut dst = ctx
+        .stream()
+        .alloc_zeros::<T>(tensor.bytes / element_bytes)?;
     {
-        let (src_ptr, _src_guard) = tensor.data.device_ptr(&ctx.stream);
-        let (dst_ptr, _dst_guard) = dst.device_ptr_mut(&ctx.stream);
+        let (src_ptr, _src_guard) = tensor.data.device_ptr(ctx.stream());
+        let (dst_ptr, _dst_guard) = dst.device_ptr_mut(ctx.stream());
         unsafe {
-            cuda_result::memcpy_dtod_async(dst_ptr, src_ptr, tensor.bytes, ctx.stream.cu_stream())
+            cuda_result::memcpy_dtod_async(dst_ptr, src_ptr, tensor.bytes, ctx.stream().cu_stream())
         }
         .with_context(|| {
             format!(
@@ -511,8 +656,8 @@ fn copy_projection_component_to_typed_contiguous<'a, T: DeviceRepr>(
             end <= expected_bytes,
             "Kimi expert-major {component} copy would exceed destination: end {end}, expected {expected_bytes}"
         );
-        let (src_ptr, _src_guard) = tensor.data.device_ptr(&ctx.stream);
-        let (dst_ptr, _dst_guard) = dst.device_ptr_mut(&ctx.stream);
+        let (src_ptr, _src_guard) = tensor.data.device_ptr(ctx.stream());
+        let (dst_ptr, _dst_guard) = dst.device_ptr_mut(ctx.stream());
         // SAFETY: this is a byte-preserving D2D copy from the raw safetensors
         // GPU payload into a typed buffer with the same total byte count. The
         // dtype and shape were validated immediately before allocation.
@@ -521,7 +666,7 @@ fn copy_projection_component_to_typed_contiguous<'a, T: DeviceRepr>(
                 dst_ptr + offset as u64,
                 src_ptr,
                 tensor.bytes,
-                ctx.stream.cu_stream(),
+                ctx.stream().cu_stream(),
             )
         }
         .with_context(|| {
